clean up padding
